pyproject.toml
[tool.poetry]
name = "mssite"
version = "0.1.0"
description = ""
authors = ["zormnus"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.2"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.14.0"
django-environ = "^0.11.2"
pillow = "^10.2.0"
python-slugify = "^8.0.4"
drf-spectacular = "^0.27.1"
drf-spectacular-sidecar = "^2024.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
